Bugfix in GeneralType.typeEquals().
Under certain circumstances every type could be considered equal to a type representing null. This leads to unexpected behavior in isomorphism and embedding checks where two nodes of different types (one of both being NULL) are matched.

Note that the previous behavior is correct in certain embedding checks: If the type of a target node is null it should be considered to match any pattern node type to avoid HRG rule duplication. The feasibility function CompatibleNodeTypes has been adapted accordingly.