include constants as external nodes in reachable fragment
this is necessary, since constants can be accessed within a method
(even if it is not "reachable" from the parameters)
and the corresponding nodes should of course match.

the test ist currently not passing, since it is already assuming
that parameters are not variables in the calling method, but this
is not yet reflected in the reachable fragment computation.