storing stateSpaceToAnalysedCall
in this case, we only care about the relation of calls to stateSpaces.
These stateSpaces could potentially be partial/uncompleted, but could
also be already complete.
Further we are in this map not interested in the states which are not
fully explored yet (when adding a stateSpace-Call pair we usually do not
know about this).
This enables us, to use PartialStateSpaces more directly to model pairs
of uncomplete stateSpaces und their containing unexplored states.