bring recursive- and nonrecursive methodExecutors together
They both have to register a dependency, as changes in the fixpointcomputation
have to be passed upwards also to non-recursive methods.
Therefore the only difference is, whether they execute the method
right away or instead register the call for later analysis.
This is realized via a template-method where the two different implementations
have to provide the strategy to generate contracts.